+++ internal/consts/consts.go
@@ -395,6 +395,12 @@ const (
 	FieldInstallationID                = "installation_id"
 	FieldAppID                         = "app_id"
 	FieldAIAPath                       = "aia_path"
+	FieldTitle                         = "title"
+	FieldMessageBase64                 = "message_base64"
+	FieldAuthenticated                 = "authenticated"
+	FieldStartTime                     = "start_time"
+	FieldEndTime                       = "end_time"
+	FieldLink                          = "link"
 
 	/*
 		common environment variables